Don't duplicate code in the README
This is bad practice as you have to change both the README and the js
file each time leading to desynchronization of both codebases.

The new link points to the file in the master branch or the repository
where it's in. So once the code is merged, the link is correct. Looking
at the js file always will give the right version of the code.